+++ src/main/scala/com/ubirch/util/ConfigBase.scala
@@ -3,5 +3,5 @@ package com.ubirch.util
 import com.typesafe.config.{ Config, ConfigFactory }
 
 trait ConfigBase {
-  def conf: Config = ConfigFactory.load()
+  lazy val conf: Config = ConfigFactory.load()
 }
