+++ src/test/scala/com/ubirch/simulations/SendAndVerifySimulation.scala
@@ -1,16 +1,11 @@
 package com.ubirch.simulations
 
 import com.ubirch.DataGenerator
-import com.ubirch.util._
 import io.gatling.core.Predef._
 
 import scala.language.postfixOps
 
-class SendAndVerifySimulation
-  extends Simulation
-  with SendAndVerifyUPP
-  with Protocols
-  with ConfigBase {
+class SendAndVerifySimulation extends Simulation with WithScenarios {
 
   val numberOfUsers: Int = conf.getInt("sendAndVerifySimulation.numberOfUsers")
   val devices: List[String] = conf.getString("simulationDevices").split(",").toList.filter(_.nonEmpty)
