+++ src/test/scala/com/ubirch/simulations/SendUPPAtOnceUserSimulation.scala
@@ -1,15 +1,10 @@
 package com.ubirch.simulations
 
-import com.ubirch.util._
 import io.gatling.core.Predef._
 
 import scala.language.postfixOps
 
-class SendUPPAtOnceUserSimulation
-  extends Simulation
-  with SendUPP
-  with Protocols
-  with ConfigBase {
+class SendUPPAtOnceUserSimulation extends Simulation with WithScenarios {
 
   val numberOfUsers: Int = conf.getInt("sendUPPAtOnceUserSimulation.numberOfUsers")
   val devices: List[String] = conf.getString("simulationDevices").split(",").toList.filter(_.nonEmpty)
