+++ src/test/scala/com/ubirch/simulations/SendUPPConstantUsersWithThrottleSimulation.scala
@@ -1,16 +1,11 @@
 package com.ubirch.simulations
 
-import com.ubirch.util.ConfigBase
 import io.gatling.core.Predef._
 
 import scala.concurrent.duration._
 import scala.language.postfixOps
 
-class SendUPPConstantUsersWithThrottleSimulation
-  extends Simulation
-  with SendUPP
-  with Protocols
-  with ConfigBase {
+class SendUPPConstantUsersWithThrottleSimulation extends Simulation with WithScenarios {
 
   // *****
   // Temporarily disabled so we can run some quick tests. //
