--- src/test/scala/com/ubirch/simulations/VerifyUPP.scala
@@ -1,24 +1,0 @@
-package com.ubirch.simulations
-
-import com.typesafe.scalalogging.LazyLogging
-import io.gatling.core.Predef._
-import io.gatling.http.Predef._
-
-import scala.language.postfixOps
-
-trait VerifyUPP extends Common with LazyLogging {
-
-  val verify = {
-    http("Verify UPP data")
-      .post("")
-      .header(HttpHeaderNames.ContentType, HttpHeaderValues.ApplicationJson)
-      .body(StringBody(session => session("HASH").as[String]))
-      .check(bodyString.find.notNull)
-
-  }
-
-  def verifyScenario(suffixes: List[String]) = {
-    getScenarioWithFileData("Device Verification (UPP)", suffixes, verify)
-  }
-
-}
