+++ src/test/scala/com/ubirch/simulations/VerifyUPPAtOnceUserSimulation.scala
@@ -1,15 +1,10 @@
 package com.ubirch.simulations
 
-import com.ubirch.util.ConfigBase
 import io.gatling.core.Predef._
 
 import scala.language.postfixOps
 
-class VerifyUPPAtOnceUserSimulation
-  extends Simulation
-  with VerifyUPP
-  with Protocols
-  with ConfigBase {
+class VerifyUPPAtOnceUserSimulation extends Simulation with WithScenarios {
 
   val numberOfUsers: Int = conf.getInt("verifyUPPAtOnceUserSimulation.numberOfUsers")
   val devices: List[String] = conf.getString("simulationDevices").split(",").toList.filter(_.nonEmpty)
